Remodel Error handling, log handling, http res & implement db error handling
We are now using erx to handle errors this commmit reflects this change
for the entire user stack. The commit also handles DB insertion and
fetch errors for users (and provides the base for other modules to error
handle) - The commit also reflects new logging policy of using proper
log levels, starting with debug when error is generated und increasing
or decreasing the severity by wrapping and stacking the error in
different layers of erx. We also introduce a cleaner and more
descriptive ways of HTTP response using utils and standard formatting
for failure and success responses - currently only implemented in users
http handler module

Signed-off-by: Sidharth Soni (Sid Sun) <sid@sidsun.com>